+++ modules/apis/src/main/java/com/axway/apim/apiimport/actions/APIQuotaManager.java
@@ -62,22 +62,14 @@ public class APIQuotaManager {
             LOG.info("{} quota for API: {} Nothing to do.", type.getFriendlyName(), createdAPI.getName());
             return;
         }
-        APIManagerAPIMethodAdapter methodAdapter = APIManagerAdapter.getInstance().getMethodAdapter();
         APIManagerQuotaAdapter quotaManager = APIManagerAdapter.getInstance().getQuotaAdapter();
         LOG.info("Updating {} quota for API: {}", type.getFriendlyName(), createdAPI.getName());
         LOG.debug("{}-Restrictions: Desired: {}, Actual: {}", type.getFriendlyName(), desiredRestrictions, actualRestrictions);
-        // In order to compare/merge the restrictions, we must translate the desired API-Method-Names, if not a "*", into the methodId of the createdAPI
-        if (desiredRestrictions != null) {
-            for (QuotaRestriction desiredRestriction : desiredRestrictions) {
-                if ("*".equals(desiredRestriction.getMethod()))
-                    continue;
-                desiredRestriction.setMethod(methodAdapter.getMethodForName(createdAPI.getId(), desiredRestriction.getMethod()).getId());
-            }
-        }
         // Load the entire current default quota
         APIQuota currentDefaultQuota = quotaManager.getDefaultQuota(type);
         LOG.debug("Current Default Quota : {}", currentDefaultQuota);
         List<QuotaRestriction> mergedRestrictions = addOrMergeRestriction(actualRestrictions, desiredRestrictions);
+        LOG.debug("Merged  Quota : {}", mergedRestrictions);
         populateMethodId(createdAPI, mergedRestrictions);
         // If there is an actual API, remove the restrictions for the current actual API
         if (actualState != null) {
@@ -132,7 +124,7 @@ public class APIQuotaManager {
             // Additionally, we have to change the methodId
             // Load the method for actualAPI to get the name of the method to which the existing quota is applied to
             if (actualState != null) {
-                APIMethod actualMethod = methodAdapter.getMethodForId(actualState.getId(), restriction.getMethod());
+                APIMethod actualMethod = methodAdapter.getMethodForName(actualState.getId(), restriction.getMethod());
                 if (actualMethod != null) {
                     // Now load the new method based on the name for the createdAPI
                     APIMethod newMethod = methodAdapter.getMethodForName(createdAPI.getId(), actualMethod.getName());
@@ -141,6 +133,10 @@ public class APIQuotaManager {
                 } else {
                     LOG.warn("API Method Name : {} not found in specification", restriction.getMethod());
                 }
+            }else {
+                // For new api creation
+                APIMethod newMethod = methodAdapter.getMethodForName(createdAPI.getId(), restriction.getMethod());
+                restriction.setMethod(newMethod.getId());
             }
         }
     }